pyproject.toml
[project]
name = "convert-mappings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "polars>=1.23.0",
    "pyyaml>=6.0.2",
]
